create helper for auth
